fix: moved benchmark_test.py into scripts dir
